Update ipaddr.js.d.ts to ipaddr.d.ts, correctly named
